Add multiprocessing to predictions.py.

+++ amazon/predictions.py
@@ -1,5 +1,7 @@
 #!/usr/local/bin/python
 
+import multiprocessing as mp
+import Queue
 from optparse import OptionParser
 import sqlite3
 import csv
@@ -13,6 +15,9 @@ from cosineSimSim import cosineSimSim
 
 # params
 outputTemplateTemplate = '%s_%%s_%%s.csv'
+workerTimeout = 15
+workerQueueSize = 100
+END_OF_QUEUE = -1
 
 # db params
 dbTimeout = 5
@@ -24,6 +29,8 @@ selectReviewsStmt =\
 
 def getParser(usage=None):
     parser = OptionParser(usage=usage)
+    parser.add_option('-w', '--numWorkers', dest='numWorkers', type='int',
+        default=1, help='Number of worker processes.', metavar='NUM')
     parser.add_option('-d', '--database', dest='db_fname',
         default='data/amazon.db', help='sqlite3 database file.', metavar='FILE')
     parser.add_option('-o', '--output-dir', dest='outputDir',
@@ -72,30 +79,68 @@ def getPredictions(dimensions, sigma, reviews1, reviews2):
             j += 1
     return predictions
 
-def processPairs(db_conn, inputfile, outputDir, outputTemplate,
-                 dimensions, sigma, cosineFunc):
+def processPair(db_conn, writer, dimensions, sigma, cosineFunc,
+                productId1, productId2):
     db_curs = db_conn.cursor()
+    # fetch product reviews
+    db_curs.execute(selectReviewsStmt, (productId1,))
+    reviews1 = [row for row in db_curs.fetchall()]
+    db_curs.execute(selectReviewsStmt, (productId2,))
+    reviews2 = [row for row in db_curs.fetchall()]
+    # compute cosine similarity using the provided function.
+    cosineSim, numUserCommon = cosineFunc(reviews1, reviews2)
+    # get predictions
+    predictions = getPredictions(dimensions, sigma, reviews1, reviews2)
+    # write output
+    for (userId, rating1, rating2, prediction) in predictions:
+        error = prediction - cosineSim
+        writer.writerow([userId, rating1, rating2, prediction, error])
+
+def worker(workerIdx, q, db_fname, outputDir, outputTemplate, randomSeed,
+           dimensions, sigma, cosineFunc):
+    num_writes = 0
+    num_skips = 0
+    # Seed random module
+    random.seed(randomSeed)
+    # connect to db
+    db_conn = sqlite3.connect(db_fname, dbTimeout)
+    while True:
+        (productId1, productId2) = q.get()
+        if productId1 == END_OF_QUEUE: break
+        outputFileName = os.path.join(outputDir, outputTemplate %
+                                      (productId1, productId2))
+        if os.path.isfile(outputFileName):
+            num_skips += 1
+            print 'Skipping %s . . .' % outputFileName
+        else:
+            print 'Writing %s . . .' % outputFileName
+            with open(outputFileName, 'wb') as csvfile:
+                writer = csv.writer(csvfile)
+                processPair(db_conn, writer, dimensions, sigma, cosineFunc,
+                            productId1, productId2)
+                num_writes += 1
+
+def master(inputfile, queues, workers):
+    i = 0
     for line in inputfile:
         tokens = line.split(',')
         productId1 = tokens[0].strip()
         productId2 = tokens[1].strip()
-        # fetch product reviews
-        db_curs.execute(selectReviewsStmt, (productId1,))
-        reviews1 = [row for row in db_curs.fetchall()]
-        db_curs.execute(selectReviewsStmt, (productId2,))
-        reviews2 = [row for row in db_curs.fetchall()]
-        # compute cosine similarity using the provided function.
-        cosineSim, numUserCommon = cosineFunc(reviews1, reviews2)
-        # get predictions
-        predictions = getPredictions(dimensions, sigma, reviews1, reviews2)
-        # write output
-        outputFileName = os.path.join(outputDir,
-            outputTemplate % (productId1, productId2))
-        with open(outputFileName, 'wb') as csvfile:
-            writer = csv.writer(csvfile)
-            for (userId, rating1, rating2, prediction) in predictions:
-                error = prediction - cosineSim
-                writer.writerow([userId, rating1, rating2, prediction, error])
+        workerIdx = i % len(workers)
+
+        while True:
+            try:
+                queues[workerIdx].put((productId1, productId2),
+                                      timeout=workerTimeout)
+                break
+            except Queue.Full:
+                print 'WARNING: Worker Queue Full!'
+        i += 1
+
+    # close queues
+    for q in queues:
+        q.put((END_OF_QUEUE, END_OF_QUEUE))
+        q.close()
 
 def main():
     # Parse options
@@ -122,21 +167,30 @@ def main():
             'Invalid Similarity function: %s' % options.cosineFunc
         return
 
-    # Seed random module
-    random.seed(options.randomSeed)
-
-    # Connect to db
-    print 'Connect to %s . . .' % options.db_fname
-    db_conn = sqlite3.connect(options.db_fname, dbTimeout)
-
-    # Open input file
-    print 'Reading from %s . . .' % inputFileName
-    inputfile = open(inputFileName, 'r')
-
-    # Make predictions
     outputTemplate = outputTemplateTemplate % options.cosineFunc
-    processPairs(db_conn, inputfile, options.outputDir, outputTemplate,
-                 options.dimensions, options.sigma, cosineFunc)
-    
+
+    # create queues
+    queues = []
+    for w in range(options.numWorkers):
+        queues.append(mp.Queue(workerQueueSize))
+
+    # create worker processes
+    workers = []
+    for w in range(options.numWorkers):
+        workers.append(mp.Process(target=worker,
+            args=(w, queues[w], options.db_fname, outputDir, outputTemplate,
+                  options.randomSeed, options.dimensions, options.sigma,
+                  cosineFunc)))
+
+    # start worker processes
+    for w in range(options.numWorkers):
+        workers[w].start()
+
+    # open input file
+    print 'Reading from %s. . .' % inputFileName
+    with open(inputFileName, 'r') as inputfile:
+        # do master task
+        master(inputfile, queues, workers)
+
 if __name__ == '__main__':
     main()
